+++ PXCTestKit/Environment.swift
@@ -12,7 +12,7 @@ final class Environment {
 
     private static let prefix = "PXCTEST_CHILD_"
 
-    static func prepare(_ environmentVariables: [String: String]?, with otherEnvironmentVariables: [String: String]) -> [String: String] {
+    static func prepare(forRunningTests environmentVariables: [String: String]?, with otherEnvironmentVariables: [String: String]) -> [String: String] {
         var result = environmentVariables ?? [:]
         for (key, value) in otherEnvironmentVariables {
             if !key.hasPrefix(prefix) {
@@ -20,6 +20,17 @@ final class Environment {
             }
             result[key.replacingOccurrences(of: prefix, with: "")] = value
         }
+        return result
+    }
+
+    static func prepare(forListingTests environmentVariables: [String: String]?) -> [String: String] {
+        var result = environmentVariables ?? [:]
+        let insertLibraries = result["DYLD_INSERT_LIBRARIES"] ?? ""
+        let listTestsShim = URL(fileURLWithPath: Bundle(for: self).bundlePath)
+            .deletingLastPathComponent()
+            .appendingPathComponent("libpxctest-list-tests.dylib")
+        assert(FileManager.default.fileExists(atPath: listTestsShim.path))
+        result["DYLD_INSERT_LIBRARIES"] = [listTestsShim.path, insertLibraries].joined(separator: ":")
         return result
     }